task: Avatar moved to base components
